Refactored user login, logout, and auth. Made middleware dir/files for auth middleware and refactored routes into controllers.
