+++ backend/src/middlewares/auth.ts
@@ -1,0 +1,30 @@
+import { Request, Response, NextFunction } from "express";
+import UserService from "../services/user-service.js";
+
+// middleware to check if user is logged in
+export const authCheck = async (
+  req: Request,
+  res: Response,
+  next: NextFunction,
+): Promise<Response | void> => {
+  try {
+    const sessionid = req.cookies.sessionid;
+    // Ensure sessionid is present and valid
+    if (
+      !sessionid ||
+      typeof sessionid !== "string" ||
+      sessionid.trim() === ""
+    ) {
+      return res.status(401).send("Unauthorized: No session ID provided");
+    }
+
+    const exists: boolean = await UserService.sessionExists(sessionid);
+    if (!exists) {
+      return res.status(401).send("Unauthorized: Invalid session ID");
+    }
+    next();
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({ error: "Internal Server Error" });
+  }
+};
