+++ backend/src/routes/vehicles-routes.ts
@@ -3,36 +3,8 @@ import VehicleController from "../controllers/vehicle-controller.js";
 const router = express.Router();
 
 // Example: Fetch all users
-router.get("/", async (req, res) => {
-  try {
-    const vehicles = await VehicleController.getAllVehicles();
-    res.json(vehicles);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Server error:" + error);
-  }
-});
+router.get("/", VehicleController.getAllVehicles);
 
-router.get("/available", async (req, res) => {
-  try {
-    const type: string = (req.query.type as string) ?? "";
-    const startDate: string = (req.query.startDate as string) ?? "";
-    const endDate: string = (req.query.endDate as string) ?? "";
-
-    const vehicles = await VehicleController.getAvailableVehicles(
-      type,
-      startDate,
-      endDate,
-    );
-    if (!vehicles) {
-      res.status(404).json({ error: "No cars found" });
-      return;
-    }
-    res.json(vehicles);
-  } catch (error) {
-    console.error(error);
-    res.status(500).send("Server error:" + error);
-  }
-});
+router.get("/available", VehicleController.getAvailableVehicles);
 
 export default router;
